spine_ai_tool/requirements.txt
pydantic
PyYAML
Pillow
loguru
rich
